Replace float with double

+++ IDMarkovChain/Algorithms/KMeans/IClusterPoint.cs
@@ -4,6 +4,6 @@ namespace IDMarkovChain.Algorithms.KMeans
     public interface IClusterPoint
     {
 
-        float Coordinate { get; }
+        double Coordinate { get; }
     }
 }
